requirements.txt
requests>=2.31.0
geopandas>=0.14.0
shapely>=2.0.0
pandas>=2.1.0
python-dotenv>=1.0.0
geojson>=3.1.0
